+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.2.3
+
+- :sparkles: Added an "autoPad" constructor to `XListTile` and `XCard` that matches the external and internal padding.
+- :recycle: Changed the example to feature internal padding on `XCard`.
+
+- :boom: Replaced `density` and `densityRatio` properties by `internalHorizontalPadding` and `internalVerticalPadding` in `xTheme`, `XListTile` and `XCard`.
+
 ## 1.2.2
 
 - :construction_worker: Switched to emojis in the CHANGELOG (matching the gitmoji.dev guide).